% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_file.R
\name{A_star}
\alias{A_star}
\title{Computes the BR adjustment}
\usage{
A_star(para, X, y, ncores, dll)
}
\arguments{
\item{para}{Parameter vector.}

\item{X}{Model matrix.}

\item{y}{Observed response.}

\item{ncores}{Number of cores used in estimation.}

\item{dll}{Name of the C++ template representing the model of interest.}
}
\value{
The returned value is a vector of the same size of \code{para}, containing the
value of the adjustment.
}
\description{
Computes the BR adjustement based on the expected Fisher information.
}
\details{
The function can be used for both the computation of the adjustment based on the empirical
approximation to the expected values and  the computation based on the Monte Carlo approximation.
In the first case, \code{X} and \code{y} would refer to observed data, while in the second
case they would refer to the simulated data. This is handled by the \code{gsolv} function.
}
\examples{
# See the vignette file
}
\references{
Kosmidis, I. and Firth, D. (2010).  A generic algorithm for reducing bias
in parametric estimation.  \emph{ Electron. J. Statist.}, 4, 1097-1112.
}
\author{
Ruggero Bellio
}
