% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_file.R
\name{exp_info1}
\alias{exp_info1}
\title{Computes the expected Fisher information}
\usage{
exp_info1(para, X, y, ncores, dll)
}
\arguments{
\item{para}{Parameter vector.}

\item{X}{Model matrix.}

\item{y}{Observed response.}

\item{ncores}{Number of cores used in estimation.}

\item{dll}{Name of the C++ template representing the model of interest.}
}
\value{
The returned value is a vector of the same size of \code{para}, containing the
value of the adjustment.
}
\description{
Computes the expected Fisher information divided by the sample size.
}
\details{
The function is just a portion of the \code{A_star} function.
}
\examples{
# See the vignette file
}
\references{
Kosmidis, I. and Firth, D. (2010).  A generic algorithm for reducing bias
in parametric estimation.  \emph{ Electron. J. Statist.}, 4, 1097-1112.
}
\author{
Ruggero Bellio
}
