% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_file.R
\name{gen_data}
\alias{gen_data}
\title{Generates a single simulated data set merging all the MC data together}
\usage{
gen_data(para, X, y, datagen, R = 500)
}
\arguments{
\item{para}{Parameter vector.}

\item{X}{Model matrix.}

\item{y}{Observed response.}

\item{datagen}{Function that generates a single dataset. It must have two arguments,
the parameter value \code{para} and the model matrix \code{X}.}

\item{R}{Size of Monte Carlo simulation. Default is 500.}
}
\description{
Generates a single simulated data set merging all the MC data together
}
