% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_sim.R
\name{genLogit}
\alias{genLogit}
\title{Generates a single simulated data set from a logistic regression model}
\usage{
genLogit(para, X)
}
\arguments{
\item{para}{Parameter vector.}

\item{X}{Model matrix.}
}
\description{
Generates a single simulated data set from a logistic regression model
}
