% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_sim.R
\name{genProbit}
\alias{genProbit}
\title{Generates a single simulated data set from a probit regression model}
\usage{
genProbit(para, X)
}
\arguments{
\item{para}{Parameter vector.}

\item{X}{Model matrix.}
}
\description{
Generates a single simulated data set from a probit regression model
}
