% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_file.R
\name{get_AD}
\alias{get_AD}
\title{Generate the AD object corresponding to the model of interest}
\usage{
get_AD(cppfile, X, y, parain)
}
\arguments{
\item{X}{Model matrix.}

\item{y}{Observed response.}

\item{parain}{Starting values (guesses) for parameter vector.}

\item{dll}{Name of a  .cpp file  containing the TMB model template.}
}
\description{
Generate the AD object corresponding to the model of interest
}
